webkit/data/layout_tests/platform/chromium-win/LayoutTests/fast/js/reparsing-semicolon-insertion-expected.txt

CONSOLE MESSAGE: line 13: Uncaught SyntaxError: Unexpected token }
FAIL successfullyParsed should be true. Threw exception ReferenceError: successfullyParsed is not defined

TEST COMPLETE

